GitHub Actions: Add script to compress image.
Using GitHub Actions, we can now compress images.
We use `image-actions` (https://github.com/calibreapp/image-actions)
to achieve the same. This helps to improve the
loading time for images.

GitHub Bot creates a new PR for the compressed
images, if there is any, whenever we push changes
to the `main` branch.

Signed-off-by: akshatdalton <akshat.dak@students.iiit.ac.in>

Fixes: #2.